requirements.txt
# requirements.txt

aiofiles==24.1.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
easyocr==1.7.2
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
fsspec==2025.9.0
fugashi==1.5.1
googletrans==4.0.2
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.6
lazy_loader==0.4
loguru==0.7.3
manga-ocr==0.1.14
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
priority==2.0.0
pyclipper==1.3.0.post6
pyperclip==1.11.0
python-bidi==0.6.6
PyYAML==6.0.3
Quart==0.20.0
redis==6.4.0
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
sniffio==1.3.1
sympy==1.14.0
termcolor==3.1.0
tifffile==2025.9.20
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.15.0
unidic-lite==1.0.8
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
